get fields of methods refactored
